Reuse indexer API types in session client & some fixes
