+++ crates/indexer/src/api/borrowers/dto.rs
@@ -1,9 +1,9 @@
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
 use crate::api::dto::AssetAmount;
 
-#[derive(Serialize, ToSchema)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
 pub struct BorrowerOverview {
     pub collateral_locked: Vec<AssetAmount>,
     pub borrowings: Vec<AssetAmount>,
