+++ crates/indexer/src/api/factories/dto.rs
@@ -1,4 +1,4 @@
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 use uuid::Uuid;
 
@@ -7,14 +7,14 @@ use simplex::simplicityhl::elements::hex::ToHex;
 use crate::api::utils::format_hex;
 use crate::models::FactoryStatus;
 
-#[derive(Serialize, PartialEq, Eq, Debug, ToSchema)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
 pub struct FactoryProgramUtxoDto {
     pub txid: String,
     pub vout: u32,
     pub created_at_height: u64,
 }
 
-#[derive(Serialize, PartialEq, Eq, Debug, ToSchema)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
 pub struct FactoryAuthUtxoDto {
     pub txid: String,
     pub vout: u32,
@@ -22,7 +22,7 @@ pub struct FactoryAuthUtxoDto {
     pub created_at_height: u64,
 }
 
-#[derive(Serialize, PartialEq, Eq, Debug, ToSchema)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
 pub struct FactoryDetailsResponse {
     pub id: Uuid,
     pub factory_asset_id: String,
