+++ crates/indexer/src/api/mod.rs
@@ -13,8 +13,15 @@ pub mod server;
 mod state;
 pub mod utils;
 
+pub use borrowers::dto::BorrowerOverview;
 pub use dto::AssetAmount;
 pub use error::*;
+pub use factories::dto::{FactoryAuthUtxoDto, FactoryDetailsResponse, FactoryProgramUtxoDto};
+pub use lenders::dto::LenderOverview;
+pub use offers::dto::{
+    OfferDetailsResponse, OfferListItemFull, OfferListItemShort, OfferListResponse, OfferUtxoDto,
+    OfferUtxoOutpointShort, OffersOverview, ParticipantDto, ParticipantShort,
+};
 pub use openapi::ApiDoc;
 pub use params::*;
 pub use state::AppState;
